Moved BeamDataset to package
